Create SoA data structures for joints
This essentially scalarizes the entire joint structure, and replaces each
scalar with a long array. There are a lot of them :(

Note: this may not be the most optimal approach from the point of TLB misses.
However, this allows us to easily SIMD the solver.

We can try converting the SoA layout to 8-at-a-time later.