Implement SoA solver
Note: this is still scalar! There is no performance benefit.

This is expected. The only thing that SoA could improve here is cache
efficiency, and existing code was very efficient from the start - for every
loop iteration it traversed the entire joint structure (or almost all of it).

However, without SoA we would not be able to use SIMD efficiently, so this is
the first step.