Reorder joints in groups for SIMD processing
We now have a big chunk of 8-joint groups followed by a small chunk of joints.

Every 8-joint group does not have any body intersections (uses 16 unique bodies)
so it can be solved in parallel.

While this results in a different joint processing order from the baseline,
there does not seem to be a good reason to prefer one order to the other.

Note that there are many possible orderings and groupings; it's possible to
split the bodies in a different way (e.g. maintaining original order better by
not using fast-remove that swaps items from the end) with a slightly more
complex implementation.