Add SSE2 solver variants
This is surprisingly easy once AVX2 work is done. The code almost works out of
the box.

SSE2 is slightly slower but not that much (30%?). The only part that does not
scale with the register width is load/store (that has to use SSE2 even in AVX2);
maybe it accounts for a significant chunk of time, or register spilling (which
may increase due to extra load/store transposes) is prevalent.

This also changes mode switching to use 'M'.