Implement Packed SoA mode
Instead of making big arrays of scalar joint data, create a struct with 8
joints inside it, laid out as SoA (every array has 8 scalars in it).

This produces a better memory access pattern - instead of 31 independent
streams we only have one, with an element size of 992 (31*8*4) bytes.

This is a bit faster as well.