Fix DenseHash performance for corner cases
DenseHash had several issues that have to do with erasing elements.

First, erasing did not update the element count; this meant that the table
that has both erase and insert activity only ever grows, rehashes get bigger
and bigger and eventually exhaust all memory.

Second, updating the element count leaves us with the tombstone problem -
if we erase all elements that we insert over and over again the table is being
filled with tombstones, which makes searching slow.

For now we're just rehashing the table when the number of tombstones exceeds
1/4th of the table size.